{"types":"C","runs":36}

+++ problem16.java
@@ -3,11 +3,11 @@ import acm.program.ConsoleProgram;
 public class problem16 extends ConsoleProgram{
 	public void run(){
 		
-		int a = readInt("Enter first number: ");
-		int b = readInt ("Enter secont number: ");
-		int c = readInt ("Enter third number: ");
+		double a = readInt("Enter first number: ");
+		double b = readInt ("Enter secont number: ");
+		double c = readInt ("Enter third number: ");
 		
-		double d = (1/(double) a) + (1/(double) b) + (1/(double) c);
+		double d = (1/a) + (1/b) + (1/c);
 		
 		println (1/d);
